Move code to AndroidLogcatPackages
