+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatPackagesWindow.cs
@@ -39,8 +39,9 @@ namespace Unity.Android.Logcat
 
     internal class AndroidLogcatPackagesWindow : EditorWindow
     {
-        MultiColumnListView m_ListView;
         AndroidLogcatRuntimeBase m_Runtime;
+        AndroidLogcatPackages m_Packages;
+        MultiColumnListView m_ListView;
         TextField m_Filter;
 
         List<PackageEntry> m_UnfilteredEntries;
@@ -57,6 +58,9 @@ namespace Unity.Android.Logcat
 
         private void OnEnable()
         {
+            if (rootVisualElement == null)
+                throw new NullReferenceException("rooVisualElement is null");
+
             m_Runtime = AndroidLogcatManager.instance.Runtime;
             m_Runtime.DeviceQuery.DevicesUpdated += UpdateEntries;
 
@@ -69,22 +73,14 @@ namespace Unity.Android.Logcat
 
             rootVisualElement.Q<TwoPaneSplitView>().RegisterCallback<GeometryChangedEvent>(InitializeLayout);
 
-            m_ListView = rootVisualElement.Q<MultiColumnListView>();
-            m_Filter = rootVisualElement.Q<TextField>("filter");
-            // TODO: take filter from settings
-            m_Filter.RegisterValueChangedCallback((s) =>
-            {
-                FilterBy(s.newValue);
-            });
-
-            m_ListView.itemsSource = m_UnfilteredEntries;
-            m_ListView.sortingEnabled = true;
-            m_ListView.columnSortingChanged += ColumnSortingChanged;
-            CreateLabel(nameof(PackageEntry.Name), (e) => e.Name);
-            CreateLabel(nameof(PackageEntry.Installer), (e) => e.Installer);
-            CreateLabel(nameof(PackageEntry.UID), (e) => e.UID);
+            m_Packages = new AndroidLogcatPackages(rootVisualElement, GetPackageEntries().ToList());
 
             rootVisualElement.Insert(0, new IMGUIContainer(DoDebuggingGUI));
+        }
+
+        private void UpdateEntries()
+        {
+            m_Packages.RefreshEntries(GetPackageEntries().ToList());
         }
 
         internal void InitializeLayout(GeometryChangedEvent e)
@@ -94,92 +90,7 @@ namespace Unity.Android.Logcat
             split.UnregisterCallback<GeometryChangedEvent>(InitializeLayout);
         }
 
-        private void ColumnSortingChanged()
-        {
-            var column = m_ListView.sortedColumns.FirstOrDefault();
-            m_FilteredEntries.Sort((x, y) =>
-            {
-                var result = 0;
-                if (column.columnName.Equals(nameof(PackageEntry.Name), StringComparison.InvariantCultureIgnoreCase))
-                {
-                    result = x.Name.CompareTo(y.Name);
-                }
-                else if (column.columnName.Equals(nameof(PackageEntry.Installer), StringComparison.InvariantCultureIgnoreCase))
-                {
-                    result = x.Installer.CompareTo(y.Installer);
-                }
-                else if (column.columnName.Equals(nameof(PackageEntry.UID), StringComparison.InvariantCultureIgnoreCase))
-                {
-                    result = x.UID.CompareTo(y.UID);
-                }
-
-                if (column.direction == SortDirection.Descending)
-                    return result;
-                return -result;
-            });
-
-            m_ListView.itemsSource = m_FilteredEntries;
-            m_ListView.RefreshItems();
-        }
-
-        private void FilterBy(string filter)
-        {
-            m_FilteredEntries.Clear();
-            if (string.IsNullOrEmpty(filter))
-            {
-                m_FilteredEntries.AddRange(m_UnfilteredEntries);
-            }
-            else
-            {
-                foreach (var e in m_UnfilteredEntries)
-                {
-                    if (e.Name.Contains(filter) ||
-                        e.Installer.Contains(filter) ||
-                        e.UID.Contains(filter))
-                        m_FilteredEntries.Add(e);
-                }
-            }
-
-            m_ListView.itemsSource = m_FilteredEntries;
-            m_ListView.RefreshItems();
-        }
-
-        private void UpdateEntries()
-        {
-            m_UnfilteredEntries = GetPackageEntries().ToList();
-            FilterBy(m_Filter.value);
-        }
-
-        void CreateLabel(string name, Func<PackageEntry, string> getText, Func<PackageEntry, string> getTooltip = null)
-        {
-            var id = name.ToLower();
-            m_ListView.columns[id].makeCell = () =>
-            {
-                var label = new PackageEntryLabel();
-                label.RegisterCallback<MouseDownEvent, PackageEntryLabel>((e, l) =>
-                {
-                    switch (e.button)
-                    {
-                        case 0:
-                            if (e.clickCount == 2)
-                            {
-                                //OnSelectEntryInListView(l.Entry);
-                            }
-                            break;
-                    }
-                }, label);
-                return label;
-            };
-
-            m_ListView.columns[id].bindCell = (element, index) =>
-            {
-                var label = GetInitializedElement<PackageEntryLabel>(element, index);
-                label.text = getText(label.Entry);
-                if (getTooltip != null)
-                    label.tooltip = getTooltip(label.Entry);
-            };
-        }
-
+        /*
         void CreateButton(string name)
         {
             var id = name.ToLower();
@@ -190,13 +101,8 @@ namespace Unity.Android.Logcat
                 button.text = "Hello";
             };
         }
+        */
 
-        T GetInitializedElement<T>(VisualElement element, int index) where T : PackageEntryVisualElement
-        {
-            var packageEntryElement = (PackageEntryVisualElement)element;
-            packageEntryElement.Entry = (PackageEntry)m_ListView.itemsSource[index];
-            return (T)packageEntryElement;
-        }
 
         PackageEntry[] GetPackageEntries()
         {
