+++ com.unity.mobile.android-logcat/Editor/UI/AndroidLogcatPackages.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 53a2b8f4589bb5149ac8ed0bbc1c48e4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
